\name{lukemlm-package}
\alias{lukemlm-package}
\alias{lukemlm}
\docType{package}
\title{
Data files for Multilevel and Longitudinal Data Analysis
}
\description{
These data files are used in the class \emph{Multilevel and Longitudinal Data Analysis}, 
offered by Douglas Luke, Ph.D.
}
\details{
\tabular{ll}{
Package: \tab lukemlm\cr
Type: \tab Package\cr
Version: \tab 0.99\cr
Date: \tab 2008-06-16\cr
License: \tab Data are for the use of the students and instructor of the above class.
}
See the individual help files for each of the data sets.
}
\author{
Douglas Luke

Maintainer: Douglas Luke <dluke@slu.edu>

}

