\name{LukeMLM-package}
\alias{LukeMLM-package}
\alias{LukeMLM}
\docType{package}
\title{
Data files for Multilevel and Longitudinal Data Analysis
}
\description{
These data files are used in the class \emph{Multilevel and Longitudinal Data Analysis}, 
offered by Douglas Luke, Ph.D.
}
\details{
\tabular{ll}{
Package: \tab LukeMLM\cr
Type: \tab Package\cr
Version: \tab 1.26\cr
Date: \tab 2015-09-28\cr
License: \tab Data are for the use of the students and instructor of the above class.
}
See the individual help files for each of the data sets.
}
\author{
Douglas Luke

Maintainer: Douglas Luke <dluke@wustl.edu>

}

